# Optimize AddCaller+AddStacktrace by sharing stacks