--- a/stacktrace.go
+++ b/stacktrace.go
@@ -37,7 +37,8 @@ var _stacktracePool = sync.Pool{
 }
 
 type stacktrace struct {
-	pcs []uintptr // program counters; always a subslice of storage
+	pcs    []uintptr // program counters; always a subslice of storage
+	frames *runtime.Frames
 
 	// The size of pcs varies depending on requirements:
 	// it will be one if the only the first frame was requested,
@@ -102,12 +103,14 @@ func captureStacktrace(skip int, depth stacktraceDepth) *stacktrace {
 		stack.pcs = stack.pcs[:numFrames]
 	}
 
+	stack.frames = runtime.CallersFrames(stack.pcs)
 	return stack
 }
 
 // Free releases resources associated with this stacktrace
 // and returns it back to the pool.
 func (st *stacktrace) Free() {
+	st.frames = nil
 	st.pcs = nil
 	_stacktracePool.Put(st)
 }
@@ -118,10 +121,10 @@ func (st *stacktrace) Count() int {
 	return len(st.pcs)
 }
 
-// First returns the first frame captured as part of the stacktrace.
-func (st *stacktrace) First() runtime.Frame {
-	frame, _ := runtime.CallersFrames(st.pcs).Next()
-	return frame
+// Next returns the next frame in the stack trace,
+// and a boolean indicating whether there are more after it.
+func (st *stacktrace) Next() (_ runtime.Frame, more bool) {
+	return st.frames.Next()
 }
 
 func takeStacktrace(skip int) string {
@@ -147,14 +150,13 @@ func newStackFormatter(b *buffer.Buffer) stackFormatter {
 	return stackFormatter{b: b}
 }
 
-// FormatStack formats all frames in the provided stacktrace -- minus
+// FormatStack formats all remaining frames in the provided stacktrace -- minus
 // the final runtime.main/runtime.goexit frame.
 func (sf *stackFormatter) FormatStack(stack *stacktrace) {
 	// Note: On the last iteration, frames.Next() returns false, with a valid
 	// frame, but we ignore this frame. The last frame is a a runtime frame which
 	// adds noise, since it's only either runtime.main or runtime.goexit.
-	frames := runtime.CallersFrames(stack.pcs)
-	for frame, more := frames.Next(); more; frame, more = frames.Next() {
+	for frame, more := stack.Next(); more; frame, more = stack.Next() {
 		sf.FormatFrame(frame)
 	}
 }