@abhinav 2022-01-26
@prashantv I've updated the numbers above after incorporating #1053.
Are we sure that's better than the version with sharing of CallerFrames?
Here is how that performs on my work laptop:

```
name                      old time/op    new time/op    delta
StackField-8                 767ns ± 1%     777ns ± 2%     ~     (p=0.083 n=8+10)
AddCallerHook-8              391ns ± 3%     390ns ± 5%     ~     (p=0.482 n=9+10)
AddCallerAndStacktrace-8     987ns ± 1%     729ns ± 1%  -26.14%  (p=0.000 n=8+10)
TakeStacktrace-8            1.13µs ± 1%    1.22µs ± 0%   +8.71%  (p=0.000 n=10+9)

name                      old alloc/op   new alloc/op   delta
StackField-8                  561B ± 0%      561B ± 0%     ~     (all equal)
AddCallerHook-8               248B ± 0%      240B ± 0%   -3.23%  (p=0.000 n=10+10)
AddCallerAndStacktrace-8      665B ± 0%      449B ± 0%  -32.48%  (p=0.000 n=10+10)
TakeStacktrace-8              496B ± 0%      496B ± 0%     ~     (all equal)

name                      old allocs/op  new allocs/op  delta
StackField-8                  3.00 ± 0%      3.00 ± 0%     ~     (all equal)
AddCallerHook-8               3.00 ± 0%      2.00 ± 0%  -33.33%  (p=0.000 n=10+10)
AddCallerAndStacktrace-8      5.00 ± 0%      3.00 ± 0%  -40.00%  (p=0.000 n=10+10)
TakeStacktrace-8              2.00 ± 0%      2.00 ± 0%     ~     (all equal)
```